==> example/5_run_Hg.R <==
# Install devtools
# install.packages("devtools")
#
# # Install Hourglass package from Github and load
# devtools::install_github("kazeera/Hourglass")
library(Hourglass)


run_from_excel("Example_IHC_sample_UserOptions - Copy.xlsx")


# scales::show_col(c("#ffd966ff", "#548135ff", "#b4a7d6ff", "#9900ffff", "#351c75ff", "#2F9E97ff", "#FFFC4Dff", "#C0632Fff", "#940e26ff", "#fae7dcff", "#4b8dbeff"))
